Fixes with read-only stubs (#215)
* Added eye icon when stub is read-only

* Added failing unit test

* When updating stub, make sure no read-only stub exists with the same ID.

* Updated UI so the "Disable stubs", "Enable stubs" and "Delete stubs" do not work with "read-only" stubs

+++ src/HttPlaceholder.Application/Stubs/Commands/UpdateStubCommand/UpdateStubCommandHandler.cs
@@ -33,6 +33,20 @@ public class UpdateStubCommandHandler : IRequestHandler<UpdateStubCommand>
             throw new ValidationException(validationResults);
         }
 
+        // Check that the stub is not read-only.
+        const string exceptionFormat = "Stub with ID '{0}' is read-only; it can not be updated through the API.";
+        var oldStub = await _stubContext.GetStubAsync(request.StubId);
+        if (oldStub?.Metadata?.ReadOnly == true)
+        {
+            throw new ValidationException(string.Format(exceptionFormat, request.StubId));
+        }
+
+        var newStub = await _stubContext.GetStubAsync(request.Stub.Id);
+        if (newStub?.Metadata?.ReadOnly == true)
+        {
+            throw new ValidationException(string.Format(exceptionFormat, request.Stub.Id));
+        }
+
         // Delete stub with same ID.
         await _stubContext.DeleteStubAsync(request.StubId);
         await _stubContext.DeleteStubAsync(request.Stub.Id);
